requirements.txt
asgiref==3.10.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
drf-yasg==1.21.7